multiagent_commute_ai/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
langchain>=0.1.0
langgraph>=0.1.0
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
pymupdf>=1.24.0
sentence-transformers>=2.3.1
scikit-learn>=1.4.0
shap>=0.44.0
numpy>=1.26.0
pandas>=2.1.0
joblib>=1.3.0
openai>=1.10.0
httpx>=0.26.0
python-multipart>=0.0.6
transformers>=4.36.0
torch>=2.1.0
